fixed shutdown method
